Remove redundant comments
After I wrote some code, I used to feed it into chatGPT for auto-commenting, this was because, to be honest, I hate commenting my code bases. But the comments just became cumbersome and to be more honest, annoying to read. So just removed the comments that I thought were unnecessary. If anybody ever ends up making a PR, I might be able add some more comments.