Fixed a bug thats been BUGGING me for a while, like my 1:04 am pun?
Basically if a title was too long it would try to center everything in history, which looked ugly. and cropped some stuff. Might make it say ... if something is too long in the future.
E.G. LINK.THAT.IS.WAYYYYYYYYYYY.TOO.LONG.TO.DISPLAY.ON.A/CHROMEBOOK
becomes:
LINK.THAT.IS.WAYYY...